Add OffsetDateTime and Instant converters
I had a converter of the type Converter<OffsetDateTime, Instant> that converted TIMESTAMP_WITH_TIMEZONE types from OffsetDateTimes to Instants. However, this resulted in a 'ClassCastException: String cannot be converted to OffsetDateTime' stemming from the 'from(OffsetDateTime)' in the converter - line 47 in AbstractAsyncQueryExecutor. This change fixes that problem and allows OffsetDateTime and Instant types to be retrieved from databases.