Do not update fields that haven't changed
Which at this point means the primary key fields. In the future, one may have a logic to detect actual changes and create records where the Pojo fields that haven't changed are not set.

This may be very useful to support MySQL/MariaDB `TIMESTAMP` fields with an `ON UPDATE CURRENT_TIMESTAMP()` feature, which are currently being overwritten by the old value in the POJO and there for are broken with vertx-jooq.